Simplify FeatureNode class
This change is based on the assumption that a feature node exists
independently of the tile to which it belongs being placed on the board
or not.

The property "turnPlaced" only makes sense after a tile is placed, and
a feature node gets new meaning after its placement.

Therefore, thinking of the graph of features on the board, a class
extending FeatureNode that embodies such property makes more sense, and
might be added soon.